change login logic to custom hook
